requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
pymysql
pyotp
qrcode[pil]
python-dotenv